refactor: consolidate tech icons, extract constants, clean up codebase
- Extract shared tech icon map to src/lib/techIcons.ts, removing duplicate
  icon/color definitions from Tag.tsx and ProjectFilter.tsx
- Extract resume PDF path to src/lib/constants.ts, used by Footer and resume page
- Standardize "Tailwind" to "Tailwind CSS" across data files
- Add missing NextAuth.js, SWR, Argos CI to icon map
- Remove unused getFeaturedProjects() from projects.ts
- Move inline styles from layout.tsx to globals.scss
- Add robots.txt and sitemap.xml
- Clean up stale BACKLOG.md entries
- Update README for GitHub profile
