+++ README.md
@@ -1,36 +1,21 @@
-This is a [Next.js](https://nextjs.org) project bootstrapped with [`create-next-app`](https://nextjs.org/docs/app/api-reference/cli/create-next-app).
+# Sarah Chow
 
-## Getting Started
+Senior Software Engineer building performant web and mobile applications with React, TypeScript, and Rust.
 
-First, run the development server:
+## Currently
 
-```bash
-npm run dev
-# or
-yarn dev
-# or
-pnpm dev
-# or
-bun dev
-```
+Building at [The HydroVac App](https://thehydrovacapp.com) — a React Native mobile platform for hydrovac service scheduling and management.
 
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+## Tech
 
-You can start editing the page by modifying `app/page.tsx`. The page auto-updates as you edit the file.
+**Languages** — JavaScript, TypeScript, Python, Rust
 
-This project uses [`next/font`](https://nextjs.org/docs/app/building-your-application/optimizing/fonts) to automatically optimize and load [Geist](https://vercel.com/font), a new font family for Vercel.
+**Frontend** — React, React Native, Next.js, Redux, D3.js
 
-## Learn More
+**Backend** — Node.js, Express, PostgreSQL, MongoDB, Firebase
 
-To learn more about Next.js, take a look at the following resources:
+**DevOps** — Docker, Kubernetes, Jest, Cypress
 
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
+## Links
 
-You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js) - your feedback and contributions are welcome!
-
-## Deploy on Vercel
-
-The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
-
-Check out our [Next.js deployment documentation](https://nextjs.org/docs/app/building-your-application/deploying) for more details.
+[xtchow.com](https://xtchow.com) | [LinkedIn](https://www.linkedin.com/in/xtchow/)
