+++ public/sitemap.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+  <url>
+    <loc>https://xtchow.com</loc>
+    <priority>1.0</priority>
+  </url>
+  <url>
+    <loc>https://xtchow.com/projects</loc>
+    <priority>0.8</priority>
+  </url>
+  <url>
+    <loc>https://xtchow.com/contact</loc>
+    <priority>0.7</priority>
+  </url>
+  <url>
+    <loc>https://xtchow.com/resume</loc>
+    <priority>0.6</priority>
+  </url>
+</urlset>
