+++ src/app/globals.scss
@@ -2,6 +2,16 @@
 @use "../styles/variables";
 @use "../styles/typography";
 
+body {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
+
+.flex-1 {
+  flex: 1;
+}
+
 // Utility classes
 .container {
   width: 100%;
