+++ src/data/projects.ts
@@ -100,7 +100,3 @@ export const projects: Project[] = [
 export function getProjectBySlug(slug: string): Project | undefined {
   return projects.find((p) => p.slug === slug);
 }
-
-export function getFeaturedProjects(): Project[] {
-  return projects.filter((p) => p.featured);
-}
