+++ src/lib/constants.ts
@@ -1,0 +1,1 @@
+export const RESUME_PATH = "/Chow, Sarah 260107.pdf";
